hello-data-deployment/docker-compose/base/jupyterhub/package_proxy/requirements.txt
aiohttp>=3.9,<4.0
pyyaml>=6.0,<7.0
packaging>=24.0,<25.0
